refactor(Day5): move to single file with Container class
